Change app package to "dev.ratajczyk.inventoryapp"
