+++ app/build.gradle
@@ -5,7 +5,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.ratajczykdev.inventoryapp"
+        applicationId "dev.ratajczyk.inventoryapp"
         minSdkVersion 23
         targetSdkVersion 28
         versionCode 1
