--- app/src/androidTest/java/com/ratajczykdev/inventoryapp/ExampleInstrumentedTest.java
@@ -1,28 +1,0 @@
-package com.ratajczykdev.inventoryapp;
-
-import android.content.Context;
-import androidx.test.InstrumentationRegistry;
-import androidx.test.runner.AndroidJUnit4;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static org.junit.Assert.*;
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest
-{
-    @Test
-    public void useAppContext()
-    {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getTargetContext();
-
-        assertEquals("com.ratajczykdev.inventoryapp", appContext.getPackageName());
-    }
-}
