--- app/src/main/java/com/ratajczykdev/inventoryapp/settings/SettingsFragment.kt
@@ -1,20 +1,0 @@
-package com.ratajczykdev.inventoryapp.settings
-
-import android.os.Bundle
-import androidx.preference.PreferenceFragmentCompat
-import com.ratajczykdev.inventoryapp.R
-
-/**
- * Loads preferences from an XML file resource
- *
- * @author Mikolaj Ratajczyk <mikolaj.ratajczyk(AT)gmail.com>
- */
-class SettingsFragment : PreferenceFragmentCompat() {
-
-    override fun onCreatePreferences(p0: Bundle?, p1: String?) {
-
-        //  Load the preferences from an XML resource
-        addPreferencesFromResource(R.xml.preferences)
-    }
-
-}
