--- app/src/main/java/com/ratajczykdev/inventoryapp/statistics/LoadingFragmentWithArgs.kt
@@ -1,17 +1,0 @@
-package com.ratajczykdev.inventoryapp.statistics
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import com.ratajczykdev.inventoryapp.MainActivity
-
-/**
- * Required for transfer data between Fragments (communicate between Fragments)
- *
- * Data flow example: [StatisticsFragment] -> [MainActivity] -> [GraphsFragment]
- *
- * @author Mikolaj Ratajczyk <mikolaj.ratajczyk(AT)gmail.com>
- */
-interface LoadingFragmentWithArgs {
-
-    fun loadFragmentWithArgs(fragment: Fragment, bundle: Bundle)
-}
