--- app/src/main/java/com/ratajczykdev/inventoryapp/statistics/StatisticsContract.kt
@@ -1,16 +1,0 @@
-package com.ratajczykdev.inventoryapp.statistics
-
-/**
- * Helper class containing constants for statistics data transfer between Activities/Fragments
- *
- * For example: Intents
- *
- * @author Mikołaj Ratajczyk <mikolaj.ratajczyk></mikolaj.ratajczyk>@gmail.com>
- */
-object StatisticsContract {
-    const val STATISTICS_MAP_NAME = "STATISTICS_MAP"
-    const val ITEMS_NUMBER_KEY = "ITEMS_NUMBER"
-    const val PRODUCTS_NUMBER_KEY = "PRODUCTS_NUMBER"
-    const val PRODUCTS_MAX_PRICE_KEY = "MAX_PRODUCTS_PRICE"
-    const val PRODUCTS_MIN_PRICE_KEY = "MIN_PRODUCTS_PRICE"
-}
