+++ app/src/main/java/dev/ratajczyk/inventoryapp/statistics/StatisticsContract.kt
@@ -1,0 +1,16 @@
+package dev.ratajczyk.inventoryapp.statistics
+
+/**
+ * Helper class containing constants for statistics data transfer between Activities/Fragments
+ *
+ * For example: Intents
+ *
+ * @author Mikołaj Ratajczyk <mikolaj.ratajczyk></mikolaj.ratajczyk>@gmail.com>
+ */
+object StatisticsContract {
+    const val STATISTICS_MAP_NAME = "STATISTICS_MAP"
+    const val ITEMS_NUMBER_KEY = "ITEMS_NUMBER"
+    const val PRODUCTS_NUMBER_KEY = "PRODUCTS_NUMBER"
+    const val PRODUCTS_MAX_PRICE_KEY = "MAX_PRODUCTS_PRICE"
+    const val PRODUCTS_MIN_PRICE_KEY = "MIN_PRODUCTS_PRICE"
+}
