+++ app/src/test/java/dev/ratajczyk/inventoryapp/tools/ImageHelperTest.java
@@ -1,0 +1,38 @@
+package dev.ratajczyk.inventoryapp.tools;
+
+import android.database.Cursor;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.MockitoJUnitRunner;
+
+import static org.junit.Assert.assertTrue;
+
+@RunWith(MockitoJUnitRunner.class)
+public class ImageHelperTest {
+
+    final static String TEXT_FROM_CURSOR_WITH_DATA = "something";
+
+    @Mock
+    private Cursor cursor;
+
+
+    @Test
+    public void getRandomUuid_noArguments_createRandomUuid() {
+        String uuid = ImageHelper.getRandomUuid();
+        assertTrue(uuid != null && !uuid.isEmpty());
+    }
+
+    @Test
+    public void getStringFromCursor_cursorWithData_notEmptyString() {
+        Mockito.when(cursor.isNull(Mockito.anyInt())).thenReturn(false);
+        Mockito.when(cursor.getString(Mockito.anyInt())).thenReturn(TEXT_FROM_CURSOR_WITH_DATA);
+
+        String receivedString = ImageHelper.getStringFromCursor(cursor, 12);
+        assertTrue(receivedString != null && !receivedString.isEmpty());
+    }
+
+
+}
